[Interaction] Greatly reduce jankiness when restoring the scroll position
Rewrote the part for restoring the scroll position after refreshing is completed. This mimics the Twitter app in this way where the scroll view is non-interactive while it slides back up from its resting position w/the refresh indicator to its resting position without it. This is feeling close to production quality.